Addressed review feedback: - changed doc url to link github issue - added return after continuation send -- oops! - use junit methods for testing message/exception matching - moved the test to the xpack ml/qa folder - added a gradle build file that appears to work - removed excess override settings from RestTest
