+++ x-pack/plugin/ml/qa/datafeed-multicluster-tests/build.gradle
@@ -1,0 +1,15 @@
+import org.elasticsearch.gradle.internal.test.RestIntegTestTask
+
+apply plugin: 'elasticsearch.internal-java-rest-test'
+apply plugin: 'elasticsearch.internal-testclusters'
+apply plugin: 'elasticsearch.rest-resources'
+
+dependencies {
+  testImplementation project(':x-pack:qa')
+  testImplementation project(path: ':test:test-clusters')
+  javaRestTestImplementation project(path: xpackModule('core'))
+}
+
+tasks.named('javaRestTest') {
+  usesDefaultDistribution("Test internally configures the clusters")
+}
